Implement WAL
Implements a simple Write-Ahead-Log.

Use io-uring to queue writes to the WAL log where needed (i.g.: not GET
commands). This shows the true power of io-uring by doing async disk io
from the main thread!

If we want to WAL but there is already a write to the WAL in progress we
store it in a backlog similar to the io-uring backlog.