Simplify buffer offset/len tracking
Instead of doing that manually in each place move it into a BufWrapView
with advanceable read - data still to be processed - and write - where
to write new data - windows.

It also makes using the buffers a lot nicer as you don't always have to
go the full rc-refcell borrow dance.

This also exposed a bug in the WAL buffer usage where we would reuse a
buffer still in WAL use.

With this out of the way we can handle reallocation next.